+++ class/actions_usernavhistory.class.php
@@ -115,6 +115,9 @@ class ActionsUserNavHistory
 
 	public function printMainArea($parameters, &$object, &$action, $hookmanager) {
 		global $user, $conf, $langs;
+		$print = GETPOST('optioncss', 'alphanohtml');
+		if($print == 'print') return 0;
+
 		$langs->load('usernavhistory@usernavhistory');
 
 		$aFilters = ['fk_user' => $user->id];